Added the fastfilereader.
The main thing of interest in here is the Mapper, which is a simple
interface to mmap a file.  It has implementations for both Unix and
Windows.  it will probably end up in EventMachine, but for now we're
sticking it with Swiftiply since it is needed by the code that handles
large file transfers.